+++ src/tweet/mod.rs
@@ -536,7 +536,7 @@ pub struct Timeline<'a> {
     ///The URL to request tweets from.
     link: &'static str,
     ///The token to authorize requests with.
-    token: &'a auth::Token<'a>,
+    token: &'a auth::Token,
     ///A handle that represents the event loop to run requests on.
     handle: &'a Handle,
     ///Optional set of params to include prior to adding lifetime navigation parameters.